requirements.txt
Flask==2.3.2
python-dotenv==1.2.1
pymongo==4.7.0
requests==2.32.5
gunicorn==20.1.0
